Rename profilerlib dir to profile
