+++ configure.ac
@@ -294,22 +294,22 @@ AC_CONFIG_LINKS([include/souffle/WriteStreamSQLite.h:src/WriteStreamSQLite.h])
 AC_CONFIG_LINKS([include/souffle/gzfstream.h:src/gzfstream.h])
 AC_CONFIG_LINKS([include/souffle/htmx86.h:src/htmx86.h])
 AC_CONFIG_LINKS([include/souffle/json11.h:src/json11.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Cell.h:src/profilerlib/Cell.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/CellInterface.h:src/profilerlib/CellInterface.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Cli.h:src/profilerlib/Cli.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/DataComparator.h:src/profilerlib/DataComparator.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Iteration.h:src/profilerlib/Iteration.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/OutputProcessor.h:src/profilerlib/OutputProcessor.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/ProgramRun.h:src/profilerlib/ProgramRun.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Reader.h:src/profilerlib/Reader.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Relation.h:src/profilerlib/Relation.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Row.h:src/profilerlib/Row.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Rule.h:src/profilerlib/Rule.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/StringUtils.h:src/profilerlib/StringUtils.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Table.h:src/profilerlib/Table.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/Tui.h:src/profilerlib/Tui.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/UserInputReader.h:src/profilerlib/UserInputReader.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/html_string.h:src/profilerlib/html_string.h])
+AC_CONFIG_LINKS([include/souffle/profile/Cell.h:src/profile/Cell.h])
+AC_CONFIG_LINKS([include/souffle/profile/CellInterface.h:src/profile/CellInterface.h])
+AC_CONFIG_LINKS([include/souffle/profile/Cli.h:src/profile/Cli.h])
+AC_CONFIG_LINKS([include/souffle/profile/DataComparator.h:src/profile/DataComparator.h])
+AC_CONFIG_LINKS([include/souffle/profile/Iteration.h:src/profile/Iteration.h])
+AC_CONFIG_LINKS([include/souffle/profile/OutputProcessor.h:src/profile/OutputProcessor.h])
+AC_CONFIG_LINKS([include/souffle/profile/ProgramRun.h:src/profile/ProgramRun.h])
+AC_CONFIG_LINKS([include/souffle/profile/Reader.h:src/profile/Reader.h])
+AC_CONFIG_LINKS([include/souffle/profile/Relation.h:src/profile/Relation.h])
+AC_CONFIG_LINKS([include/souffle/profile/Row.h:src/profile/Row.h])
+AC_CONFIG_LINKS([include/souffle/profile/Rule.h:src/profile/Rule.h])
+AC_CONFIG_LINKS([include/souffle/profile/StringUtils.h:src/profile/StringUtils.h])
+AC_CONFIG_LINKS([include/souffle/profile/Table.h:src/profile/Table.h])
+AC_CONFIG_LINKS([include/souffle/profile/Tui.h:src/profile/Tui.h])
+AC_CONFIG_LINKS([include/souffle/profile/UserInputReader.h:src/profile/UserInputReader.h])
+AC_CONFIG_LINKS([include/souffle/profile/html_string.h:src/profile/html_string.h])
 
 AM_MISSING_PROG([AUTOM4TE], [autom4te])
