+++ src/profile/gui_src/chartist-plugin-tooltip.js
