+++ src/profile/gui_src/chartiststyle.css
